+++ react/components/Input/index.js
@@ -129,13 +129,13 @@ class Input extends Component {
     const borderRadius = `${borderRadiusBase} ${
       prefix ? 'bl-0 br--right ' : ''
     } ${suffix ? 'br-0 br--left ' : ''}`
-    let prefixClasses = `${borderRadiusBase} br-0 br--left `
-    let suffixClasses = `${borderRadiusBase} bl-0 br--right `
+    let prefixClasses = `vtex-input__prefix ${borderRadiusBase} br-0 br--left `
+    let suffixClasses = `vtex-input__suffix ${borderRadiusBase} bl-0 br--right `
 
     let classes = `${styles.input} ${widthClass} ${box} ${styles.hideDecorators} ${styles.noAppearance} ${borderRadius} bn outline-0 `
     let labelClasses = 'vtex-input__label db mb3 w-100 c-on-base '
     let prefixAndSuffixClasses =
-      'vtex-input__prefix c-muted-2 fw5 flex items-center c-muted-2 t-body '
+      'c-muted-2 fw5 flex items-center c-muted-2 t-body '
     let prefixSuffixGroupClasses =
       'vtex-input-prefix__group flex flex-row items-stretch '
     prefixSuffixGroupClasses += `${borderRadiusBase} ${borderBase} ${