+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - `Table`'s `headerRenderer` can be used in columns that are either fixed or sortable
+### Fixed
+
+- Broken styles of `Dropdown` inside an `Input` prefix or suffix.
 
 ## [9.112.15] - 2020-03-10